requirements.txt
PySide6>=6.5.0
SQLAlchemy>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
wxauto>=1.0.0


python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
pywin32>=305 